move test_api helpers to generic helper module to avoid
circular imports
