+++ tests/functional/utils/generic.py
@@ -14,6 +14,8 @@ from contextlib import contextmanager
 from pprint import pprint
 
 from collections import namedtuple
+from directory_api_client.testapiclient import DirectoryTestAPIClient
+from directory_sso_api_client.testapiclient import DirectorySSOTestAPIClient
 from random import choice
 from string import ascii_uppercase
 from typing import List
@@ -37,12 +39,7 @@ from tests.functional.utils.context_utils import (
     Feedback,
     Message
 )
-from tests.functional.utils.test_api_utils import (
-    get_company_email,
-    delete_supplier_data_from_sso,
-    delete_supplier_data_from_dir
-)
-from tests.functional.utils.request import Method, make_request
+from tests.functional.utils.request import Method, make_request, check_response
 from tests.settings import (
     MAILGUN_DIRECTORY_API_USER,
     MAILGUN_DIRECTORY_EVENTS_URL,
@@ -52,7 +49,18 @@ from tests.settings import (
     TEST_IMAGES_DIR,
     JPEGs,
     JPGs,
-    PNGs
+    PNGs,
+    DIRECTORY_API_URL,
+    DIRECTORY_API_CLIENT_KEY,
+    SSO_PROXY_API_CLIENT_BASE_URL,
+    SSO_PROXY_SIGNATURE_SECRET
+)
+
+
+DIRECTORY_CLIENT = DirectoryTestAPIClient(
+    DIRECTORY_API_URL, DIRECTORY_API_CLIENT_KEY)
+SSO_CLIENT = DirectorySSOTestAPIClient(
+    SSO_PROXY_API_CLIENT_BASE_URL, SSO_PROXY_SIGNATURE_SECRET
 )
 
 # a type hint for a List of Company named tuples
@@ -1092,3 +1100,103 @@ def get_number_of_search_result_pages(response: Response) -> int:
     page_numbers = [int(word) for word in pages.split() if word.isdigit()]
     last_page = page_numbers[-1] if len(page_numbers) == 2 else no_page_counter
     return last_page
+
+
+def get_company_email(number: str) -> str:
+    """Get email address associated with company."""
+    response = DIRECTORY_CLIENT.get_company_by_ch_id(number)
+    check_response(response, 200)
+    email = response.json()['company_email']
+    logging.debug("Email for company %s is %s", number, email)
+    return email
+
+
+def get_published_companies(context: Context) -> list:
+    """Get a List of dicts with published companies.
+
+    :return: a list of dictionaries with published companies
+    """
+    response = DIRECTORY_CLIENT.get_published_companies()
+    context.response = response
+    check_response(response, 200)
+    return response.json()
+
+
+def get_published_companies_with_n_sectors(
+        context: Context, number_of_sectors: int) -> list:
+    """Get a List of published companies with at least N associated sectors.
+
+    :return: a list of dictionaries with matching published companies
+    """
+    response = DIRECTORY_CLIENT.get_published_companies(
+        minimal_number_of_sectors=number_of_sectors)
+    context.response = response
+    check_response(response, 200)
+    return response.json()
+
+
+@retry(wait_fixed=1000, stop_max_attempt_number=10)
+def get_verification_code(context: Context, company_number: str):
+    """Will get the verification code (sent by post) for specified company.
+
+    :return: verification code sent by post
+    """
+    print("GETTING VERIFICATION CODE")
+    response = DIRECTORY_CLIENT.get_company_by_ch_id(company_number)
+    context.response = response
+    if response.status_code != 200:
+        print("GOT {} instead of 200".format(response.status_code))
+        print(response.request.url)
+        print(response.request.headers)
+        print(response.content)
+    check_response(response, 200)
+    verification_code = response.json()['letter_verification_code']
+    return verification_code
+
+
+def is_verification_letter_sent(
+        context: Context, company_number: str) -> bool:
+    """Check if verification letter was sent.
+
+    :return: True if letter was sent and False if it wasn't
+    """
+    response = DIRECTORY_CLIENT.get_company_by_ch_id(company_number)
+    context.response = response
+    check_response(response, 200)
+    result = response.json()['is_verification_letter_sent']
+    return result
+
+
+def delete_supplier_data_from_sso(
+        email_address: str, *, context: Context = None):
+    response = SSO_CLIENT.delete_user_by_email(email_address)
+    if context:
+        context.response = response
+    if response.status_code == 204:
+        logging.debug(
+            "Successfully deleted %s user data from SSO DB", email_address)
+    else:
+        logging.error(
+            "Something went wrong when trying to delete user data for %s from "
+            "SSO DB", email_address)
+
+
+def delete_supplier_data_from_dir(ch_id: str, *, context: Context = None):
+    response = DIRECTORY_CLIENT.delete_company_by_ch_id(ch_id)
+    if context:
+        context.response = response
+    if response.status_code == 204:
+        logging.debug(
+            "Successfully deleted supplier data for company %s from DIR DB",
+            ch_id)
+    else:
+        logging.error(
+            "Something went wrong when trying to delete supplier data for "
+            "company %s from DIR DB", ch_id)
+
+
+def flag_sso_account_as_verified(context: Context, email_address: str):
+    response = SSO_CLIENT.flag_user_email_as_verified_or_not(
+        email_address, verified=True)
+    context.response = response
+    check_response(response, 204)
